% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porto_data.R
\docType{data}
\name{fire01meantemp}
\alias{fire01meantemp}
\title{Fire and extreme temperature in portugal}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(porto)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\references{
Tilloy et al. (2020)
}
\keyword{datasets}
