% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement.funct.R
\name{AnalogSel}
\alias{AnalogSel}
\title{Automatically select analaugous datasets from the 60 datasets created in Tilloy et al.(2020)
see: https://nhess.copernicus.org/articles/20/2091/2020/nhess-20-2091-2020.html for more detaila}
\usage{
AnalogSel(u2)
}
\arguments{
\item{u2}{Two column data frame}
}
\value{
No return value, called for side effects
}
\description{
Automatically select analaugous datasets from the 60 datasets created in Tilloy et al.(2020)
see: https://nhess.copernicus.org/articles/20/2091/2020/nhess-20-2091-2020.html for more detaila
}
\examples{
data(porto)
AnalogSel(fire01meantemp)
}
